refactor(client): separate theme from App.tsx
