+++ packages/client/src/theme.ts
@@ -1,0 +1,36 @@
+import { createTheme } from '@mui/material';
+
+const theme = createTheme({
+  palette: {
+    primary: {
+      main: '#1b5e20',
+    },
+    secondary: {
+      main: '#cddc39',
+    },
+    background: {
+      default: '#F4F3DD',
+    },
+  },
+  typography: {
+    fontFamily: 'Noto Sans',
+    fontSize: 14,
+    h1: {
+      fontWeight: 400,
+    },
+    h2: {
+      fontWeight: 400,
+    },
+  },
+  components: {
+    MuiAppBar: {
+      styleOverrides: {
+        colorPrimary: {
+          backgroundColor: '#1B5E2014',
+        },
+      },
+    },
+  },
+});
+
+export { theme };
